+++ CHANGELOG.md
@@ -4,5 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Syntax highlighting for content ([#3](https://github.com/scm-manager/scm-content-search-plugin/pull/3))
+
 ## 0.9 - 21-09-08
 - Initial implementation of the content search